Roles

+++ 3D-models/pages/productos/index.php
@@ -141,8 +141,9 @@ $resultado = mysqli_query($enlace, $consulta);
                 <a href="../auth/login.php" class="btn btn-success mb-2">🛒</a>
               <?php endif; ?>
 
-              
-              <?php if (isset($_SESSION['usuario']) && $_SESSION['usuario']['tipo_usuario'] === 'vendedor'): ?>
+               
+
+              <?php if (isset($_SESSION['usuario']) && $_SESSION['usuario']['tipo_usuario'] === 'vendedor' && $_SESSION['usuario']['id'] === $id_producto = $prod['id']): ?>
                 <a href="edit.php?id=<?php echo $prod['id']; ?>" class="btn btn-success mb-2">Editar</a>
                 <a href="delete.php?id=<?php echo $prod['id']; ?>" class="btn btn-danger mb-2" onclick="return confirm('¿Seguro querés eliminar este producto?');">Eliminar</a>
               <?php endif; ?>
